Crate nautilus_core

PoseiTrader is an open-source, high-performance, production-grade algorithmic trading platform, providing quantitative traders with the ability to backtest portfolios of automated trading strategies on historical data with an event-driven engine, and also deploy those same strategies live, with no code changes.

Feature flags

This crate provides feature flags to control source code inclusion during compilation, depending on the intended use case, i.e. whether to provide Python bindings for the posei_trader Python package, or as part of a Rust only build.

ffi: Enables the C foreign function interface (FFI) from cbindgen.

python: Enables Python bindings from PyO3.

Re-exports

pub use crate::nanos::UnixNanos;

pub use crate::time::AtomicTime;

pub use crate::uuid::UUID4;

Modules

collections

Abstraction layer over common hash-based containers.

consts

Core constants.

correctness

Functions for correctness checks similar to the design by contract philosophy.

datetime

Common data and time functions.

env

ffi

C foreign function interface (FFI) from cbindgen.

math

message

Common message types.

nanos

A UnixNanos type for working with timestamps in nanoseconds since the UNIX epoch.

parsing

Core parsing functions.

paths

Utility functions for resolving project and workspace directory paths.

python

Python bindings from PyO3.

serialization

Common serialization traits and functions.

time

The core AtomicTime for real-time and static clocks.

uuid

A UUID4 Universally Unique Identifier (UUID) version 4 (RFC 4122).