nautilus_core::python

Trait IntoPyObjectPoseiExt

pub trait IntoPyObjectPoseiExt<'py>: IntoPyObjectExt<'py> {
    // Provided method
    fn into_py_any_unwrap(self, py: Python<'py>) -> PyObject { ... }
}

Extend IntoPyObjectExt helper trait to unwrap PyObject after conversion.

Provided Methods

fn into_py_any_unwrap(self, py: Python<'py>) -> PyObject

Convert self into a [PyObject] while panicking if the conversion fails.

This is a convenience wrapper around [IntoPyObjectExt::into_py_any] that avoids the cumbersome Result handling when we are certain that the conversion cannot fail (for instance when we are converting primitives or other types that already implement the necessary PyO3 traits).

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.