Crate nautilus_core

NautilusTrader is an open-source, high-performance, production-grade algorithmic trading platform, providing quantitative traders with the ability to backtest portfolios of automated trading strategies on historical data with an event-driven engine, and also deploy those same strategies live, with no code changes.

Feature flags

This crate provides feature flags to control source code inclusion during compilation, depending on the intended use case, i.e. whether to provide Python bindings for the main nautilus_trader Python package, or as part of a Rust only build.

ffi: Enables the C foreign function interface (FFI) from cbindgen

python: Enables Python bindings from pyo3

Modules

correctness

Provides static condition checks similar to the design by contract philosophy to help ensure logical correctness.

datetime

Provides common data and time functions.

equality

ffi

Provides a C foreign function interface (FFI) from cbindgen.

message

Defines common message types.

nanos

Defines UnixNanos type for working with UNIX epoch (nanoseconds).

parsing

Provides core parsing functions.

python

Python bindings from pyo3.

serialization

Defines common serialization traits.

time

Provides the core AtomicTime real-time and static clocks.

uuid

A core UUID4 universally unique identifier (UUID) version 4 based on a 128-bit label (RFC 4122).