Module nautilus_core::ffi::string

Functions

cstr_drop⚠

Drops the C string memory at the pointer.

cstr_to_bytes⚠

Convert a C string pointer into bytes.

cstr_to_str⚠

Convert a C string pointer into a static string slice.

cstr_to_ustr⚠

Convert a C string pointer into an owned String.

optional_cstr_to_str⚠

Convert a C string pointer into an owned Option<String>.

optional_cstr_to_ustr⚠

Convert a C string pointer into an owned Option<Ustr>.

pystr_to_string⚠

Returns an owned string from a valid Python object pointer.

str_to_cstr

Create a C string pointer to newly allocated memory from a &str.